Add convert trading UI and admin-configurable pair catalog
